requirements.txt
Flask
flask-restx
flask-sqlalchemy
flask-migrate
alembic
pytest
requests
